Authorization -finished

+++ backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post("",checkAuth,multer({ storage: storage }).single("image"),
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -63,15 +64,23 @@ router.put("/:id",checkAuth,multer({ storage: storage }).single("image") ,
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath:imagePath
+    imagePath:imagePath,
+    creator: req.userData.userId
   });
 
-
-  Post.updateOne({ _id: req.params.id }, post)
+  // to check Authorization nModified will be 0 if the user is unAuthorized
+  Post.updateOne({ _id: req.params.id ,creator:req.userData.userId }, post)
     .then((result) => {
-      res.status(200).json({
-        message: 'update successful'
-      });
+      if(result.nModified >0){
+        res.status(200).json({
+          message: 'update successful'
+        });
+      } else {
+        res.status(401).json({
+          message: 'Not Authorized!'
+        });
+
+      }
     });
 });
 
@@ -119,14 +128,19 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete("/:id",checkAuth, (req, res, next) => {
-  Post.deleteOne({
-    _id: req.params.id
-  })
+  Post.deleteOne({_id: req.params.id , creator:req.userData.userId})
     .then(result => {
-      //  console.log(result);
-      res.status(200).json({
-        message: 'Post deleted!'
-      })
+
+      if(result.deletednCount >0){
+        res.status(200).json({
+          message: 'Deletion succeful successful'
+        });
+      } else {
+        res.status(401).json({
+          message: 'Not Authorized!'
+        });
+
+      }
     })
 });
